#31 Renamed files relevant to styling the product page for more consistency  throughout different folders.
